move study to study app, and add logical funds transfers
