+++ geeksurvey/templates/home.html
@@ -38,13 +38,14 @@
         <div class="container px-5">
             <div class="row gx-5 justify-content-center">
                 <div class="col-lg-6">
+                    <img class="mx-auto d-block" src="{% static 'gkslogo.png' %}" width="200px" height="200px">
                     <div class="text-center my-5">
                         <h1 class="display-5 fw-bolder text-white mb-2">Revolutionizing Software Engineering Research</h1>
                         <p class="lead text-white-50 mb-4">
                           Conduct and participate in software engineering research studies with GeekSurvey, a platform that solves the researcher-participant coordination problem!
                         </p>
                         <div class="d-grid gap-3 d-sm-flex justify-content-sm-center">
-                            <a class="btn btn-primary btn-lg px-4 me-sm-3" href="{% url 'account_signup' %}">Get Started</a>
+                            <a class="btn btn-success btn-lg px-4 me-sm-3" href="{% url 'account_signup' %}">Get Started</a>
                             <a class="btn btn-outline-light btn-lg px-4" href="#features">Learn More</a>
                         </div>
                     </div>
