+++ geeksurvey/templates/navbar.html
@@ -6,7 +6,7 @@
 <!-- Responsive Navigation Bar -->
 <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
   <div class="container px-5">
-      <img src="{% static 'gkslogo.png' %}" width="50px" height="50px">
+      <img src="{% static 'gkslogo_dark.png' %}" width="50px" height="50px">
       <a class="navbar-brand " href="{% url 'home' %}">GeekSurvey</a>
       <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent"
         aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
