--- geeksurvey/templates/paypal/failure.html
@@ -1,7 +1,0 @@
-
-{% extends "base.html" %}
-{% block content %}
-
-<h1>PAYPAL FAILURE</h1>
-
-{% endblock %}
