--- geeksurvey/templates/paypal/success.html
@@ -1,7 +1,0 @@
-
-{% extends "base.html" %}
-{% block content %}
-
-<h1>PAYPAL SUCCESS</h1>
-
-{% endblock %}
