--- geeksurvey/templates/profile/fund.html
@@ -1,14 +1,0 @@
-{% extends "base.html" %}
-
-
-{% block content %}
-
-<h1>Fund Profile</h1>
-<hr>
-<!-- These are somewhat arbitrary values, determined for software proof-of-concept rather than business purposes -->
-<p> Pay $125 to add $100 to your account </p>
-
-{{form.render}}
-
-{% endblock %}
-
