+++ geeksurvey/templates/research/index.html
@@ -17,9 +17,12 @@
       <img src="{% static 'books.png' %}" class="card-img-top" />
       <div class="card-body">
         <h5 class="card-title">{{study.title}}</h5>
-        <p class="card-text">{{ study.description}} </p>
+        <p class="card-text">{{study.description}} </p>
+        <p class="card-text">${{study.balance}} </p>
         <a href="{% url 'study_landing_page' study.id %}" class="btn btn-secondary">View</a>
         <a href="{% url 'study_edit' study.id %}" class="btn btn-primary">Edit</a>
+        <br>
+        <a style="margin-top:5px;" href="{% url 'study_funds' study.id %}" class="btn btn-success">Manage Funds</a>
       </div>
     </div>
