+++ geeksurvey/urls.py
@@ -27,17 +27,13 @@ urlpatterns = [
     path('accounts/', include('allauth.urls')),
     path('admin/', admin.site.urls),
     path('payments/', include('payments.urls')),
+    path('study/', include('study.urls')),
     path('paypal/', include('paypal.standard.ipn.urls'), name='paypal-ipn'),
     path('participate/', views.participate, name='participate'),
     path('participate/find/', views.part_discover, name='part_discover'),
-    path('study/<uuid:study_id>/', views.study_landing_page, name='study_landing_page'),
-    path('study/enroll/<uuid:study_id>/', views.study_enroll, name='study_enroll'),
-    path('study/enroll/<uuid:study_id>/fail/', views.study_enroll_fail, name='study_enroll_fail'),
-    path('study/complete/<uuid:study_id>/', views.study_complete, name='study_complete'),
     path('profile/', views.profile, name='profile'),
     path('profile/view/<str:username>', views.profile_view, name='profile_view'),
     path('profile/update/', views.profile_update, name='profile_update'),
     path('research/', views.research, name='research'),
-    path('study/edit/<uuid:study_id>/', views.study_edit, name='study_edit'),
-    path('study/create/', views.study_create, name='study_create'),
+
 ]
