+++ payments/README.md
@@ -1,0 +1,3 @@
+The payments app contains
+all external payments logic
+
