+++ study/README.md
@@ -1,0 +1,6 @@
+This app contains
+study creation,
+study editing,
+study balance,
+enrollment,
+completion
