+++ study/templates/study/complete.html
@@ -1,0 +1,20 @@
+{% extends "base.html" %}
+
+{% load static %}
+{% load crispy_forms_tags %}
+
+{% block content %}
+  <h1> {{study.title}} </h1>
+  <h3> Completion Page </h3>
+  <p> Please complete the <a href="{{study.survey_url}}">external survey</a> and then enter the completion code here </p>
+
+  <form method="POST" enctype="multipart/form-data">
+  {% csrf_token %}
+  <fieldset class="form-group">
+      {{ complete_form|crispy }} <!-- Study Completion Form -->
+  </fieldset>
+  <div class="form-group">
+      <button class="btn btn-outline-info" type="submit">Submit</button>
+  </div>
+
+{% endblock %}
