+++ study/templates/study/landing.html
@@ -1,0 +1,47 @@
+{% extends "base.html" %}
+
+{% load static %}
+
+{% block content %}
+  <h1> {{study.title}} </h1>
+  {% if request.user in study.enrolled.all %}
+    <span class="badge badge-success">Enrolled</span>
+  {% endif %}
+  {% if request.user in study.completed.all %}
+    <span class="badge badge-primary">Completed</span>
+  {% endif %}
+  <br>
+  <img src="{% static 'books.png' %}" width="100px" />
+  <h5> {{study.description}} </h5>
+  <p> Created By: <a href="{% url 'profile_view' study.owner %}"> {{study.owner}} </a> </p>
+
+  <hr>
+  <p> Enrolled:
+    {% for enrolled_user in study.enrolled.all %}
+    <a href="{% url 'profile_view' enrolled_user %}"> {{ enrolled_user }} </a>,
+    {% endfor %}
+  </p>
+  <p> Completed:
+    {% for completed_user in study.completed.all %}
+    <a href="{% url 'profile_view' completed_user %}"> {{ completed_user }} </a>,
+    {% endfor %}
+  </p>
+  <hr>
+
+  {% if not request.user in study.enrolled.all %}
+    <form action="{% url 'study_enroll' study.id %}" method="POST">
+      {% csrf_token %}
+      <input type="submit" value="Enroll" />
+    </form>
+  {% else %}
+    {% if not request.user in study.completed.all %}
+      <span class="badge badge-warning">UNCOMPLETED</span>
+      <br>
+      <h4>You've enrolled in this study, but you still need to complete it.</h4>
+ 
+      <a href="{% url 'study_complete' study.id %}">Complete this study</a>
+    {% endif %}
+  {% endif %}
+
+{% endblock %}
+
